revert rpn_matching

+++ tools/rpn_matching.py
@@ -47,12 +47,12 @@ def get_cdf(dets_nms_all, CONF_THRESH=np.linspace(0,1,11)):
         num_rois = np.sum(dets_nms_all[:, 4] >= conf)
         cdf = cdf.append(gl.SFrame({"conf": [conf], "num_rois": [num_rois]})) 
     return cdf
-def transform_and_build_nn(cand_sf, dfe, db="./features_sframe.gl", radius=0.6, k=1):   
+def transform_and_build_nn(cand_sf, dfe, db="./features_sframe.gl", radius=0.51, k=1):   
     cand_sf = dfe.transform(cand_sf)
     cand_sf = cand_sf.add_row_number()
     db_sf = gl.SFrame(db)
     db_sf = db_sf.add_row_number()
-    nn = gl.nearest_neighbors.create(db_sf,features=['deep_features.image'],distance='cosine')
+    nn = gl.nearest_neighbors.create(db_sf,label="pid", features=['deep_features.image'],distance='cosine')
     neighbors = nn.query(cand_sf,radius=radius,k=k)
     return neighbors, db_sf, cand_sf
 
@@ -80,10 +80,10 @@ def demo(net, image_name, db="./features_sframe.gl", NMS_THRESH_GLOBAL=0.6):
 
     # Load the demo image
     #im_file = os.path.join(cfg.DATA_DIR, 'demo', image_name)
-    if isStringType(image_name):
+    if image_name.isinstance(basestring):
       im_file = os.path.join(cfg.DATA_DIR, 'demo', 'imgs', image_name)
       im = cv2.imread(im_file)
-    if image_name is gl.Image:
+    else:
       im = gl.Image._to_pil_image(image_name)
 
     # Detect all object classes and regress object bounds
@@ -119,9 +119,9 @@ def demo(net, image_name, db="./features_sframe.gl", NMS_THRESH_GLOBAL=0.6):
     #dfe = gl.feature_engineering.DeepFeatureExtractor('image', model='auto', output_column_prefix=feat)
     dfe = gl.load_model('./alexnet.gl')
     # 28 imgs in the catalogue, calculates c(100)*n(28) = 2800 similarities
-    neighbors, db_sf, cand_sf = transform_and_build_nn(rois_sf, dfe, db=db, .6, 2)
-    neighbors
-    db_sf.remove_column('path')
+    neighbors, db_sf, cand_sf = transform_and_build_nn(rois_sf, dfe, db=db, .6, 1)
+    if "path" in db_sf.column_names()
+      db_sf.remove_column('path')
     return neighbors, db_sf, cand_sf
 
 def parse_args():
